src/backend/pyproject.toml
#
# Meet package
#
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "meet"
version = "0.1.33"
authors = [{ "name" = "DINUM", "email" = "dev@mail.numerique.gouv.fr" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django",
    "Framework :: Django :: 5",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
description = "A simple video and phone conferencing tool, powered by LiveKit"
keywords = ["Django", "Contacts", "Templates", "RBAC"]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3==1.38.42",
    "Brotli==1.1.0",
    "brevo-python==1.1.2",
    "celery[redis]==5.5.3",
    "django-configurations==2.5.1",
    "django-cors-headers==4.7.0",
    "django-countries==7.6.1",
    "django-lasuite[all]==0.0.10",
    "django-parler==2.3",
    "redis==5.2.1",
    "django-redis==6.0.0",
    "django-storages[s3]==1.14.6",
    "django-timezone-field>=5.1",
    "django==5.2.3",
    "djangorestframework==3.16.0",
    "drf_spectacular==0.28.0",
    "dockerflow==2024.4.2",
    "easy_thumbnails==2.10",
    "factory_boy==3.3.3",
    "gunicorn==23.0.0",
    "jsonschema==4.24.0",
    "markdown==3.8.2",
    "nested-multipart-parser==1.5.0",
    "psycopg[binary]==3.2.9",
    "PyJWT==2.10.1",
    "python-frontmatter==1.1.0",
    "requests==2.32.4",
    "sentry-sdk==2.30.0",
    "whitenoise==6.9.0",
    "mozilla-django-oidc==4.0.1",
    "livekit-api==1.0.4",
    "livekit-agents==1.2.6",
    "livekit-plugins-gladia==1.0.23",
    "aiohttp==3.12.13",
]

[project.urls]
"Bug Tracker" = "https://github.com/numerique-gouv/meet/issues/new"
"Changelog" = "https://github.com/numerique-gouv/meet/blob/main/CHANGELOG.md"
"Homepage" = "https://github.com/numerique-gouv/meet"
"Repository" = "https://github.com/numerique-gouv/meet"

[project.optional-dependencies]
dev = [
    "django-extensions==4.1",
    "drf-spectacular-sidecar==2025.6.1",
    "freezegun==1.5.2",
    "ipdb==0.13.13",
    "ipython==9.3.0",
    "pyfakefs==5.9.1",
    "pylint-django==2.6.1",
    "pylint==3.3.7",
    "pytest-cov==6.2.1",
    "pytest-django==4.11.1",
    "pytest==8.4.1",
    "pytest-icdiff==0.9",
    "pytest-xdist==3.7.0",
    "responses==0.25.7",
    "ruff==0.12.0",
    "types-requests==2.32.4.20250611",
]

[tool.setuptools]
packages = { find = { where = ["."], exclude = ["tests"] } }
zip-safe = true

[tool.distutils.bdist_wheel]
universal = true

[tool.ruff]
exclude = [
    ".git",
    ".venv",
    "build",
    "venv",
    "__pycache__",
    "*/migrations/*",
]
line-length = 88


[tool.ruff.lint]
select = [
    "B", # flake8-bugbear
    "BLE", # flake8-blind-except
    "C4", # flake8-comprehensions
    "DJ", # flake8-django
    "I", # isort
    "PLC", # pylint-convention
    "PLE", # pylint-error
    "PLR", # pylint-refactoring
    "PLW", # pylint-warning
    "RUF100", # Ruff unused-noqa
    "RUF200", # Ruff check pyproject.toml
    "S", # flake8-bandit
    "SLF", # flake8-self
    "T20", # flake8-print
]
ignore= ["DJ001", "PLR2004"]

[tool.ruff.lint.isort]
section-order = ["future","standard-library","django","third-party","meet","first-party","local-folder"]
sections = { meet=["core"], django=["django"] }

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["S", "SLF"]

[tool.pytest.ini_options]
addopts = [
    "-v",
    "--cov-report",
    "term-missing",
    # Allow test files to have the same name in different directories.
    "--import-mode=importlib",
]
python_files = [
    "test_*.py",
    "tests.py",
]
